Update LICENSE
The content of this page is investment, It shall not be treated as marketing advice or independent research. INCOMEFLOW shall not be responsible for any direct or indirect loss resulting from the use of the provided information. Investing in financial involves risk so I suggest as an investor,you should invest what you can afford to lose, and there is guarantee investors will get their Profit and capital invested and Past performance does not guarantee future returns.

INCOMEFLOW (registration number 40203015744, legal address: Roberta Hirsa iela 1, Riga, LV-1045, Latvia) is an investment firm licenced and supervised by the central bank of Latvia. License number 27-55/2023/2. Financial instruments distributed by incomeflow and legal relations arising from them are governed by the law of the Republic of Latvia. The place of provision of incomeflow services is the Republic of Latvia, regardless of the investors' country of origin and residence.

INCOMEFLOW is a member of the national investor compensation scheme established under Directive 97/9/EC. The scheme protects investors by providing compensation if incomeflow fails to return financial instruments or cash to investors. The maximum compensation an investor can claim under the scheme is 90% of their net loss, up to a maximum of EUR 20 000.